+++ lib/buttons-and-links/KButton.vue
@@ -22,11 +22,12 @@
       data-test="iconBefore"
       class="prop-icon"
     />
-
-    <!-- @slot Pass sub-components into the button, which provides more flexibility than and takes precedence over the `text` prop -->
+    <!-- @slot Pass sub-components into the button, typically `KDropdownMenu` -->
+    <slot name="menu"></slot>
+    <!-- @slot Slot alternative to the `text` prop -->
     <slot v-if="$slots.default"></slot>
 
-    <template v-else>
+    <template>
       <span class="link-text" :style="textStyle">{{ text }}</span>
     </template>