+++ lib/buttons-and-links/KButton.vue
@@ -22,13 +22,12 @@
       data-test="iconBefore"
       class="prop-icon"
     />
+    <!-- @slot Pass sub-components into the button, typically `KDropdownMenu` -->
+    <slot name="menu"></slot>
 
-    <!-- @slot Pass sub-components into the button, which provides more flexibility than and takes precedence over the `text` prop -->
     <slot v-if="$slots.default"></slot>
 
-    <template v-else>
-      <span class="link-text" :style="textStyle">{{ text }}</span>
-    </template>
+    <span class="link-text" :style="textStyle">{{ text }}</span>
 
     <!-- @slot Slot alternative to the `iconAfter` prop -->
     <slot name="iconAfter"></slot>
@@ -89,8 +88,7 @@
         default: 'button',
       },
       /**
-       * @ignore
-       * Adds a dropdown arrow - internal use only
+       * Adds a dropdown arrow
        */
       hasDropdown: {
         type: Boolean,